requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.7.14
click==8.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
mcp==1.12.3
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
referencing==0.36.2
rpds-py==0.26.0
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
Pillow>=10.0.0